rgw: fix rgw crash in fips mode.
aws s3 etags are supposed to be md5 checksums of the object contents.
It is not possible to calculate these without md5.  openssl provides
an exception mechanism to access non-fips algorithms when necessary,
and this appears to me to be such a case.

Note: while most of the rgw code use of md5 seems to be etags,
there are other uses as well, and I don't have any information
on the expected behavior for those cases.

Resolves: rhbz#1789237
Signed-off-by: Marcus Watts <mwatts@redhat.com>