mon: elector: Fix a connectivity-mode first-election bug
commit e7996f16bee1b14982b41efcbfab06780bbf65fb
("mon: Guard new elector pinging and election strategies behind a mon_feature")
added backward compatibility by preventing us from sending pings
until we've formed a quorum of mons that understand pings. But
forming a quorum when you start in connectivity mode requires non-zero
scores, which you don't get by looking up un-pinged peers. So elections failed!

The root cause here is that I overloaded "0" as a score that means
"not allowed to be leader" in addition to "is not trusted". That was
foolish, so stop overloading it. Instead, allow monitors to become
leaders with a 0 score; have connectivity_election_score return -1
for disallowed leaders; and initialize leader_score to -1 as well
so that if we haven't acked a leader yet, we'll do SOMETHING useful
on propose.

This means monitors can win elections before pings complete, which
in one of our unit tests results in them winning when some of the
losers realize they have better scores. That triggers our
victory_makes_sense() check, so adjust it to simply bump the epoch
and elect again when the check fails.

Signed-off-by: Greg Farnum <gfarnum@redhat.com>
(cherry picked from commit 0541a5acb71369925b05afbba7b55c12de236f37)

Resolves: rhbz#1800382