librbd: helper state machine for computing diffs between object-maps
This is a port of the existing synchronous logic from the diff iterate
API but converted to be asynchrous. The goal is to re-use this logic
when performing a deep-copy image sync to skip any objects that have
not been updated.

Signed-off-by: Jason Dillaman <dillaman@redhat.com>
(cherry picked from commit 330f2a7bb94fff4555280e7fc5ad85dbdcfaaaea)

Resolves: rhbz#1860073