mon/OSDMonitor: only take in osd into consideration when trimming osdmaps
we should not take down osd into consideration when trimming osdmap. in
e62269c892, we decrease the upper bound of range of osdmaps to be trimmed
if the given osd is out. but we should have to decrease it only if the
osd in question is still *in*.

so, in this change, the min_lec is decreased only if the osd in question
is *in*.

Fixes: https://tracker.ceph.com/issues/47290
Signed-off-by: Kefu Chai <kchai@redhat.com>
(cherry picked from commit 79175740cf2394bba74163ca8a5131419b9641ed)

 Conflicts:
	src/mon/OSDMonitor.cc - trivial resolution

(cherry picked from commit 50be1fc6040cf647484a47e2d3620fc758fd6d05)

Resolves: rhbz#1875628