qa/tasks/openssl_keys.py: add subjectAltName to certificates
Get rid of this annoying teuthology log message which appears
many many times:

.../urllib3/connection.py:395: SubjectAltNameWarning: Certificate
for <some_host> has no `subjectAltName`, falling back to check for a
`commonName` for now. This feature is being removed by major browsers and
deprecated by RFC 2818. (See https://github.com/urllib3/urllib3/issues/497
for details.)

I'm also adding the ip address, which also allows https://IPaddress/
This is part of the standard and works with most clients, but some
versions of python ignore this.  C'est la vie.

I've included some cleverness in the logic to only add SAN when
adding a certificate for a host (and not some other principal).

Fixes: https://tracker.ceph.com/issues/48177
Signed-off-by: Marcus Watts <mwatts@redhat.com>