pybind/cephfs: use stub implementations when building document
also define a stub for rados.Rados right in cephfs.pyx, so the cython
compiler needs it. but it's cubersome to copy rados.pxd from rados
python binding when building cephfs python binding. because, if we
install a local python package using requirement.txt, pip does not allow
us to specify the include-directory for Cython. and cephfs/setup.py is
located out of the source tree when it is being built by pip, and unlike
in our CMake based build, which specifies --cython-include-dirs with
absolute directory, we don't have access to the root directory of
project in Read the Docs environment, so it's impossible for us to
locate pybind/rados in setup.py.

we could pass "--global-option" as part of package specifier in
requirements.txt, like:

cephfs --global-option="build_ext" --global-option="--for-doc"

or

cephfs --global-option="build_doc"

but then we will have to override or create a command
for interpreting the command line options.

Signed-off-by: Kefu Chai <kchai@redhat.com>