osd: Add mclock profile infrastructure and implement mclock profiles
Define config options to set max osd capacity (hdd & ssd) to be allocated
between clients of dmclock namely,
 - "osd_mclock_max_capacity_iops"
 - "osd_mclock_max_capacity_iops_hdd"
 - "osd_mclock_max_capacity_iops_ssd"

Define config option "osd_mclock_profile" to specify the buit-in profile
to enable.

Also, Set the number of op shards being used in the osd within the mclock
scheduler as well. This is necessary to calculate the per shard limits
within the mclock scheduler.

With the above information, enable the specified mclock profile by
calling the appropriate method to set the profile specific mclock
parameters and Ceph options.

Signed-off-by: Sridhar Seshasayee <sseshasa@redhat.com>