osd: Fix the OpSchedulerItem cost scaling calculation.
Calculate the scaled cost for an OpSchedulerItem in msec based on the
overall osd capacity (across all shards). Remove logic to cache the
cost of the previous op in order to re-use the cost for the next
transaction (if unchanged) as it doesn't seem to provide any benefit in
terms of performance.

Signed-off-by: Sridhar Seshasayee <sseshasa@redhat.com>