rgw: register daemon in service map with more details
From now on, the rgw daemon will register in the service map with more
details. We have added the zonegroup name, the zone name and the RADOS
client ID to the rgw daemon. This will give us better visibility when
query the Ceph status and also will also us to run multiple rgws with
the same cephX user AND to be displayed individually. Previously, using a
single cephX user to run multiple gateways was possible but the Ceph
status would display a single rgw.
Also, in the same situation it will allow us to gather metrics and
export them through the ceph-mgr Prometheus exporter.

Fixes: https://tracker.ceph.com/issues/49227
Signed-off-by: Sébastien Han <seb@redhat.com>
Signed-off-by: Jiffin Tony Thottan <jthottan@redhat.com>