mon/CreatingPGs: do not compatible with pre-octopus versions
creating_pgs_t is only used by monitor to persist the pg being created,
so a quincy monitor is not likely to see a creating_pgs_t encoded by
a pre-octopus monitor as long as user does not upgrade a cluster across
two releases when creating a new pool.

in octopus, pool_create_info is encoded with version 3.

Signed-off-by: Kefu Chai <kchai@redhat.com>