cephadm: fix escaping/quoting of stderr-prefix arg for ceph daemons
On some versions of podman doing

   --default-log-stderr-prefix="debug "

seems to work okay, but on others it does not, producing lines like

   /usr/bin/ceph-mon: "debug "2021-02-26T15:30:36.266+0000 7f3d8a955700  4 rocksdb:                    Options.db_write_buffer_size: 0

Escaping/quoting is wonky because we are producing a bit of bash for the
unit.run file.  Quote the entire arguemnt when it has a space, which is
sufficient for this case.

Signed-off-by: Sage Weil <sage@newdream.net>