mgr/cephadm: rgw: do not mess with realm configuration
It is simpler to consider this out of scope for the orchestrator.  The
user should set up their multisite realms/zones before deploying the
daemons (or the daemons will not start).  In the future we can wrap this
with a more friendly tool, perhaps.

Signed-off-by: Sage Weil <sage@newdream.net>
(cherry picked from commit 4498f7dc204cada316ba0ea73ca9954d23c437d2)