auth/CephxClientHandler: always process provided service tickets
Since nautilus, CEPHX_GET_AUTH_SESSION_KEY shares both the auth ticket
and the service tickets which spares an extra round-trip to get the
service tickets via CEPHX_GET_PRINCIPAL_SESSION_KEY.  This applies to
msgr1 as well, but we don't take advantage of it on the client side.
Because in msgr1 case session_key and connection_secret are NULL,
extra_tickets bufferlist is just skipped and the client ends up sending
a separate CEPHX_GET_PRINCIPAL_SESSION_KEY request as if talking to
a pre-nautilus monitor.

Signed-off-by: Ilya Dryomov <idryomov@gmail.com>