mon: fail fast when unauthorized global_id (re)use is disallowed
When unauthorized global_id (re)use is disallowed, we don't want to
let unpatched clients in because they wouldn't be able to reestablish
their monitor session later, resulting in subtle hangs and disrupted
user workloads.

Denying the initial connect for all legacy (CephXAuthenticate < v3)
clients is not feasible because a large subset of them never stopped
presenting their ticket on reconnects and are therefore compatible with
enforcing mode: most notably all kernel clients but also pre-luminous
userspace clients.  They don't need to be patched and excluding them
would significantly hamper the adoption of enforcing mode.

Instead, force clients that we are not sure about to reconnect shortly
after they go through authentication and obtain global_id.  This is
done in Monitor::dispatch_op() to capture both msgr1 and msgr2, most
likely instead of sharing a cluster map.  The client would come back
(perhaps to a different monitor) and attempt to reclaim:

- for msgr2, the client is kicked before the monmap is shared.  This
  means that the client is still in MonClient::authenticate() and the
  reconnect error would be appropriately propagated to the user, same
  as a regular authentication error (e.g. for a bad key).

- for msgr1, the monmap is sent immediately after the final MAuthReply.
  When the client is kicked, it will have already gotten the monmap and
  moved on from MonClient::authenticate() considering authentication to
  be finished.  The reconnect error would not propagate and the client
  would hang waiting for remaining cluster maps.  This is not ideal but
  acceptable because since there is no proper higher-level session at
  that point, the user doesn't get a chance to start a workload.

To help with identifying clients that need patching, add global_id and
global_id_status to to "sessions" output.

Signed-off-by: Ilya Dryomov <idryomov@gmail.com>